requirements.txt
streamlit
pandas
numpy
pandas-ta
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl